restructure for new runtime
